requirements.txt
fastapi
uvicorn
sqlalchemy
spotipy
python-dotenv
pytest
pytest-asyncio
httpx
psycopg2-binary